remove unused rust code
